Create ExtractWordsUtil
