+++ package.json
@@ -8,7 +8,6 @@
     "eslint": "^0.24.1",
     "rework": "^1.0.1",
     "uglifyjs": "^2.4.10",
-    "underscore": "^1.8.3",
     "yargs": "^3.10.0"
   },
   "devDependencies": {
