--- src/ContentSelectorExtraction.js
@@ -1,56 +1,0 @@
-var _ = require('underscore');
-
-var ContentSelectorExtraction = function (content) {
-  this.content = content;
-
-  // Get hash of word-parts appearing in content
-  this.contentWords = getAllUsedWords(content);
-};
-
-module.exports = ContentSelectorExtraction;
-
-ContentSelectorExtraction.prototype.filter = function (words) {
-  var that = this;
-
-  return _.filter(words, function (word) {
-    var splitWord = word.toLowerCase().split('-');
-
-    if (splitWord.length === 1) {
-      return that.contentWords[word];
-    } else {
-      return that.contentWords[word] ||
-             _.every(splitWord, function (part) {
-               return that.contentWords[part] || that.contentWords[part + '-'] ||
-                      that.contentWords['-' + part + '-'] || that.contentWords['-' + part];
-             });
-    }
-  });
-};
-
-ContentSelectorExtraction.prototype.filterBySearch = function (words) {
-  var that = this;
-
-  return _.filter(words, function (word) {
-    return that.content.indexOf(word.replace(/\\/g, '')) > -1;
-  });
-};
-
-var getAllUsedWords = function (content) {
-  var used = {};
-  var word = '';
-
-  for (var i = 0; i < content.length; i++) {
-    var chr = content[i];
-
-    if (chr.match(/[a-z]+/)) {
-      word += chr;
-    } else {
-      used[word] = true;
-      word = '';
-    }
-  }
-
-  used[word] = true;
-
-  return used;
-};
