+++ src/CssTreeWalker.js
@@ -1,8 +1,8 @@
 var EventEmitter = require('events').EventEmitter;
 var rework = require('rework');
 
-var ReworkTypes = require('./constants/ReworkTypes');
-var EventTypes = require('./constants/EventTypes');
+var RULE_TYPE = 'rule';
+var MEDIA_TYPE = 'media';
 
 var CssTreeWalker = function (code, plugins) {
   EventEmitter.call(this);
@@ -24,18 +24,18 @@ CssTreeWalker.prototype.beginReading = function () {
 
 CssTreeWalker.prototype.readPlugin = function (tree) {
   this.readRules(tree.rules);
-  this.finishReading(tree.rules);
+  this.removeEmptyRules(tree.rules);
 };
 
 CssTreeWalker.prototype.readRules = function (rules) {
   rules.forEach(function (rule) {
     var ruleType = rule.type;
 
-    if (ruleType === ReworkTypes.RULE_TYPE) {
-      this.emit(EventTypes.READ_SELECTOR, rule.selectors, rule);
+    if (ruleType === RULE_TYPE) {
+      this.emit('readRule', rule.selectors, rule);
     }
 
-    if (ruleType === ReworkTypes.MEDIA_TYPE) {
+    if (ruleType === MEDIA_TYPE) {
       this.readRules(rule.rules);
     }
   }.bind(this));
@@ -49,8 +49,28 @@ CssTreeWalker.prototype.toString = function () {
   return '';
 };
 
-CssTreeWalker.prototype.finishReading = function (rules) {
-  this.emit(EventTypes.FINISH_READING, rules);
+CssTreeWalker.prototype.removeEmptyRules = function (rules) {
+  var emptyRules = [];
+
+  rules.forEach(function (rule) {
+    var ruleType = rule.type;
+
+    if (ruleType === RULE_TYPE && rule.selectors.length === 0) {
+      emptyRules.push(rule);
+    }
+
+    if (ruleType === MEDIA_TYPE) {
+      this.removeEmptyRules(rule.rules);
+      if (rule.rules.length === 0) {
+        emptyRules.push(rule);
+      }
+    }
+  }.bind(this));
+
+  emptyRules.forEach(function (emptyRule) {
+    var index = rules.indexOf(emptyRule);
+    rules.splice(index, 1);
+  });
 };
 
 module.exports = CssTreeWalker;
