--- src/constants/EventTypes.js
@@ -1,6 +1,0 @@
-var EventTypes = {
-  READ_SELECTOR: 'readSelectors',
-  FINISH_READING: 'finishedReading'
-};
-
-module.exports = EventTypes;
