--- src/constants/ReworkTypes.js
@@ -1,6 +1,0 @@
-var ReworkTypes = {
-  RULE_TYPE: 'rule',
-  MEDIA_TYPE: 'media'
-};
-
-module.exports = ReworkTypes;
