+++ src/utils/PrintUtil.js
@@ -1,35 +1,9 @@
-var _ = require('underscore');
-
-var getRuleString = function (twig) {
-  var ruleString = '';
-  for (var i = 1; i < twig.length; i++) {
-    var rulePart = twig[i];
-    switch (rulePart[0]) {
-    case 's':
-      ruleString += rulePart[1] !== '\n' ? rulePart[1] : '';
-      break;
-    case 'clazz':
-      ruleString += '.' + rulePart[1][1];
-      break;
-    case 'shash':
-      ruleString += '#' + rulePart[1];
-      break;
-    case 'ident':
-      ruleString += rulePart[1];
-      break;
-    case 'attrib':
-      ruleString += '[' + rulePart[1][1] + ']';
-      break;
-    default:
-      ruleString += 'Unsupported: ' + JSON.stringify(twig);
-    }
-  }
-  return ruleString;
-};
+var startTime;
+var beginningLength;
 
 var PrintUtil = {
 
-  printInfo: function (startTime, beginningLength, endingLength) {
+  printInfo: function (endingLength) {
     var logFn = console.error;
     logFn.call(null, '##################################');
     logFn.call(null, 'Before purify, CSS was ' + beginningLength + ' chars long.');
@@ -43,8 +17,13 @@ var PrintUtil = {
     var logFn = console.error;
     logFn.call(null, '##################################');
     logFn.call(null, 'Rejected selectors:');
-    logFn.call(null, _.map(rejectedTwigs, getRuleString).join('\n'));
+    logFn.call(null, rejectedTwigs.join('\n'));
     logFn.call(null, '##################################');
+  },
+
+  startLog: function (cssLength) {
+    startTime = new Date();
+    beginningLength = cssLength;
   }
 
 };
